rotate-api-tokens.sh: add new commands change-ft-account-passwords and sync-ft-account-passwords
change-ft-account-passwords will "detect" email/password combinations in a
set of FT variables by looking for matching pairs of settings ending in
_email and _password. it will then generate a new password for each
account, separate accounts being determined by unique email address. a
github PR will then be created to apply these changes.

sync-ft-account-passwords will use the API to set the account passwords
for any accounts found in a set of FT variables (detected in the same way).

doing it this way allows separate accounts to be used for the different
test variants, new classes of accounts could be added, or some accounts
could be omitted entirely for certain variants and the scripts should
continue to operate robustly & agnostically

neither of these scripts are particularly fast because i didn't really want
to stash an unencrypted copy of this data in e.g. a tmpfile anywhere
and so we keep referring back to the encrypted file, needing new key
negotiation etc.