filter out e-signature suppliers without an agreementID
In the case where a supplier had not signed the agreement and was on an e-signature framework we were erroneously looking up their agreement which would then 404 and cause an exception on the API client. This filters out suppliers without an agreement ID so that subsequent steps are not run for thm.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>