Get list of closed briefs more efficiently
Instead of getting all the closed briefs and then filtering by date, use the options built into the API to do the filtering in the API. This gets us the list of briefs closed on that date much faster.

Since we're no longer doing our own filtering, we no longer need the tests.